Add files via upload
Split the color write out to 3 individual writes instead of a single multi-register write. It's less efficient, but the goal is to send more traffic that can be monitored by the firewall for demo purposes. I also added error handling and connection reset when a write command isn't successful (because we're blocking a specific register value on the firewall). It makes the app behave more gracefully during a demo. Now it will just chug along waiting for a new color after a few seconds rather than crashing the MODBus connection completely requiring an app restart.